Fixed Odom Publisher.

+++ base_controller/src/main/java/com/github/c77/base_driver/OdometryStatus.java
@@ -4,9 +4,9 @@ package com.github.c77.base_driver;
  * @author jcerruti@willowgarage.com (Julian Cerruti)
  */
 public interface OdometryStatus {
-    int getPoseX();
+    double getPoseX();
 
-    int getPoseY();
+    double getPoseY();
 
     double getPoseTheta();
